update Graphic11 Etc

+++ Framework11/ui/dockwindow.cpp
@@ -147,15 +147,15 @@ bool cDockWindow::Undock(const bool newWindow // = true
 	{
 		ReleaseCapture();
 
-		const int width = 800;
-		const int height = 600;
+		const float width = 800;
+		const float height = 600;
 		cImGui *oldGui = &m_owner->m_gui;
 		cRenderWindow *window = cDockManager::Get()->NewRenderWindow(m_name, width, height);
 		window->m_dock = this;
 		m_parent = NULL;
 		SendMessage(m_owner->getSystemHandle(), WM_LBUTTONUP, 0, 0); // bug fix (ImGui::ItemActive() bug)
 		m_owner = window;
-		CalcResizeWindow(eDockResize::RENDER_WINDOW, sRectf(0, 0, (float)width, (float)height));
+		CalcResizeWindow(eDockResize::RENDER_WINDOW, sRectf(0, 0, width, height));
 		oldGui->SetContext();
 		window->requestFocus();
 		window->SetDragState();
